plugin/skills/microsoft-foundry/foundry-agent/invoke/scripts/requirements.txt
azure-ai-projects
azure-identity
requests
